flow: replace old logging package with v2 logging package
